Level-Beginner/Next_Word_Predictor/requirements.txt.txt
flask==3.0.0
numpy==1.26.4
tensorflow==2.15.0
keras==2.15.0
pyspellchecker==0.7.2